requirements.txt
psycopg2
python-dotenv~=1.0.1
Django~=5.1.4
djangorestframework~=3.15.2